# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.User:
    properties:
      avatar:
        description: 头像
        type: string
      created_at:
        type: string
      email:
        description: 邮箱
        type: string
      id:
        type: integer
      nickname:
        description: 昵称
        type: string
      phone:
        description: 手机号
        type: string
      status:
        description: '1: 正常, 2: 禁用'
        type: integer
      updated_at:
        type: string
    type: object
  errors.AppError:
    properties:
      code:
        description: 错误码
        type: integer
      data:
        description: 返回数据
      message:
        description: 错误信息
        type: string
    type: object
...
  handler.LoginRequest:
    properties:
      account:
        example: "13800138000"
        type: string
      credential:
        example: "123456"
        type: string
      login_type:
        enum:
        - phone_password
        - email_password
        - phone_code
        example: phone_password
        type: string
    required:
    - account
    - credential
    - login_type
    type: object
  handler.LoginResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  handler.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: "123456"
        minLength: 6
        type: string
      phone:
        example: "13800138000"
        type: string
    required:
    - password
    - phone
    type: object
  handler.SystemStatus:
    properties:
      arch:
        description: 系统架构
        type: string
      cpu_usage:
        description: CPU使用率
        type: number
      disk_total:
        description: 总磁盘空间（字节）
        type: integer
      disk_used:
        description: 已用磁盘空间（字节）
        type: integer
      memory_total:
        description: 总内存（字节）
        type: integer
      memory_used:
        description: 已用内存（字节）
        type: integer
      os:
        description: 操作系统
        type: string
      uptime:
        description: 系统运行时间（秒）
        type: integer
    type: object
...
host: localhost:11223
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: MC Dashboard 后端服务API文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MC Dashboard API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: 用户登录接口，支持手机号密码、邮箱密码、手机验证码登录
      parameters:
      - description: 登录信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.AppError'
      summary: 用户登录
      tags:
      - 用户管理
  /register:
    post:
      consumes:
      - application/json
      description: 用户注册接口
      parameters:
      - description: 注册信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.AppError'
      summary: 用户注册
      tags:
      - 用户管理
...
  /system/status:
    get:
      consumes:
      - application/json
      description: 获取当前系统的运行状态，包括CPU、内存、磁盘使用情况等
      produces:
      - application/json
      responses:
        "200":
          description: 系统状态信息
          schema:
            $ref: '#/definitions/handler.SystemStatus'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - Bearer: []
      summary: 获取系统状态
      tags:
      - 系统
  /user:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - Bearer: []
      summary: 获取用户信息
      tags:
      - 用户管理
  /user/logout:
    post:
      consumes:
      - application/json
      description: 用户登出接口，清除用户登录状态
      produces:
      - application/json
      responses:
        "200":
          description: 登出成功
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - Bearer: []
      summary: 用户登出
      tags:
      - 用户管理
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: 请输入Bearer token，格式为：Bearer <token>
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"